Clean up web3 creation
The reason behind this change is to make cljs-web3-next the central place to
obtain the web3 instance. This library will load the Web3.js (has it as
a dependency) and others who want to use it only need to provide info
about the provider (how to connect to the network).

This can be done via
1. providing URL of the node to connect to
2. passing the provider instance (e.g. `window.ethereum` after authorization)