Remove legacy code for in-browser Web3 initialization
Removed bunch of code that dealt with the `window.web3` (that MetaMask
used to inject). Now we assume that `window.ethereum` is the way to get
the provider in UI for Web3.js use, when there's user interaction (in
browser for example).

Also this change adapts to the `cljs-web3-next` as the central place to
obtain Web3 instance. Other higher-level libraries don't have to require
Web3.js directly or call `new Web3`. Instead they give configuration
specific to their needs (window.ethereum or url) and ask `cljs-web3-next` to
give them the web3 instance